{{cookiecutter.app_name}}/requirements.txt
fastapi[all]
motor
python-jose[cryptography]
passlib[bcrypt]
python-dotenv
strawberry-graphql[fastapi]